Move OS definitions from mod package to os
